feat/API15 Update plugin for API15

+++ DalamudPlugin/RaidsRewritten/EncounterManager.cs
@@ -73,7 +73,7 @@ public sealed class EncounterManager(
         "vfx/common/eff/dk04ht_win0h.avfx",
         "vfx/common/eff/cmat_aoz0f.avfx",
     ];
-    private readonly Dictionary<ushort, IEncounter> encounterMap = encounters.AsValueEnumerable().ToDictionary(e => e.TerritoryId, e => e);
+    private readonly Dictionary<uint, IEncounter> encounterMap = encounters.AsValueEnumerable().ToDictionary(e => e.TerritoryId, e => e);
 
     private bool? inCombat = null;
     private byte weather = 0;
@@ -105,7 +105,7 @@ public sealed class EncounterManager(
         dalamud.Framework.Update -= OnFrameworkUpdate;
     }
 
-    private void OnTerritoryChanged(ushort obj)
+    private void OnTerritoryChanged(uint obj)
     {
         ActiveEncounter?.Unload();
         statusPartyListProcessor.Reset();
@@ -135,7 +135,7 @@ public sealed class EncounterManager(
         if (configuration.EverythingDisabled) { return; }
     }
 
-    private void OnObjectEffect(uint Target, ushort Param1, ushort Param2)
+    private void OnObjectEffect(uint Target, uint Param1, uint Param2)
     {
         var gameObject = dalamud.ObjectTable.SearchByEntityId(Target);
         if (gameObject == null) { return; }
